Make Calculators list readonly and refactor MakeCalculator function
The Calculators list in CalcFactory is now a readonly field. This is to ensure safety as the list does not change after initialization. The MakeCalculator function has been refactored for readability, increasing the maintainability of the code. It now uses a foreach loop, instead of FirstOrDefault method, to find and return the desired calculator. If no match is found, it will provide the first defined calculator as a default.