Refactor item selection in FixCommonErrors form
The logic for manipulating the status of list items has been refactored to reduce code duplication. The previously separate methods for selecting all items and inversing the selection are now utilising a new method UpdateItemsStatus(). This method takes in a status evaluator function to determine the new status of each item, increasing code reusability and enhancing readability.