Refactor and optimize ChangeCasingNames in subtitle form
Removed private field _usedNames and the method AddToListViewNames as they were no longer needed after refactoring. The method FindAllNames was refactored and optimized to improve performance and readability. HashSet for foundNameLookup was introduced to store all the found names in the subtitle text. Addition of the all found names from the subtitle text has been simplified with ListView item creation. Also the method IsEmbeddedWord was added for better clarity and maintainability. It checks whether the name from the name list is embedded in the main text.