Refactor HttpClient creation and usage
The commit introduces a refactoring of HttpClient creation and usage. The previous factory method "MakeHttpClient()" is replaced with "Create()" in the DownloaderFactory and HttpClientFactory and applied throughout the code. Also, HttpClientFactory has been implemented to handle the creation of HttpClient instances with customizable proxy settings. This will ensure more consistent and manageable code.

Signed-off-by: Ivandro Jao <ivandrofly@gmail.com>